misc/shreksy/chal/requirements.txt
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
fonttools==4.54.1
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mypy-extensions==1.0.0
numpy<2
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.2
pycryptodome==3.20.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
